bug fixes
- temp offset is now float
- moved temp upper and lower limits into if else blocks

+++ RPi_firmware/TRV_v6.2_scripts/auto_scripts/pid_temp.py
@@ -135,7 +135,7 @@ def update_PID_params():
         sample_time = int(config[0]['acf']['pid_sample_time'])
         pid_lower_limit = int(config[0]['acf']['pid_lower_limit'])
         pid_upper_limit = int(config[0]['acf']['pid_upper_limit'])
-        pid_temp_offset = int(config[0]['acf']['pid_temp_offset'])
+        pid_temp_offset = float(config[0]['acf']['pid_temp_offset'])
         pid_temp_sleep_interval = int(config[0]['acf']['pid_temp_sleep_interval'])
 
         pid.tunings = (kp, ki, kd)
